+++ src/Xteq5Launcher/Program.cs
@@ -41,10 +41,6 @@ namespace Xteq5Launcher
             }
 
 
-            //Old code
-            //Application.EnableVisualStyles();
-            //Application.SetCompatibleTextRenderingDefault(false);
-            //Application.Run(new Form1());
 
         }
     }